requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
PyMySQL==1.1.0
pyodbc==5.1.0
pytz==2024.1
spyne==2.14.0
suds==1.1.2
Werkzeug==3.0.1
